+++ src/capacitor-sync.ts
@@ -2,6 +2,7 @@ import { Project } from './project';
 import { MonoRepoType } from './monorepo';
 import { exists } from './analyzer';
 import { InternalCommand } from './command-name';
+import { preflightNPMCheck } from './node-commands';
 
 /**
  * Creates the capacitor sync command
@@ -9,15 +10,17 @@ import { InternalCommand } from './command-name';
  * @returns string
  */
 export function capacitorSync(project: Project): string {
+  const preop = preflightNPMCheck(project);
+
   const ionicCLI = exists('@ionic/cli');
   switch (project.repoType) {
     case MonoRepoType.none:
-      return ionicCLI ? capCLISync() : ionicCLISync();
+      return preop + (ionicCLI ? capCLISync() : ionicCLISync());
     case MonoRepoType.folder:
     case MonoRepoType.npm:
-      return InternalCommand.cwd + (ionicCLI ? capCLISync() : ionicCLISync());
+      return InternalCommand.cwd + preop + (ionicCLI ? capCLISync() : ionicCLISync());
     case MonoRepoType.nx:
-      return nxSync(project);
+      return preop + nxSync(project);
     default:
       throw new Error('Unsupported Monorepo type');
   }